Add simplified config flow for testing dependency installation
Temporarily replace complex config_flow.py with minimal version that:
- Only imports const.py (no heavy dependencies)
- Has basic user input form
- Creates config entry to trigger integration setup
- Avoids importing gateway.py, config_helpers.py, schema.py

This tests if the 'Invalid handler specified' error was caused by
early dependency loading during config flow discovery phase.

The simplified flow should allow:
1. Config flow to load without dependency errors
2. Dependencies to be installed during async_setup_entry
3. Integration to initialize properly

Files:
- config_flow.py -> config_flow_full.py (backup)
- config_flow_simple.py -> config_flow.py (active)
