+++ custom_components/eltako/config_flow.py
@@ -1,9 +1,6 @@
-"""Config flows for the Eltako integration."""
-# https://developers.home-assistant.io/docs/config_entries_config_flow_handler
-
+"""Simplified config flow for testing dependency installation."""
 from __future__ import annotations
 
-import ipaddress
 import logging
 from typing import Any
 
@@ -12,202 +9,76 @@ import voluptuous as vol
 from homeassistant import config_entries
 from homeassistant.core import callback
 from homeassistant.data_entry_flow import FlowResult
-from homeassistant.helpers import device_registry as dr
 
-from . import gateway
-from . import config_helpers
-from .const import *
-from .schema import CONFIG_SCHEMA
+from .const import DOMAIN, CONF_GATEWAY_DESCRIPTION, CONF_SERIAL_PATH
 
 _LOGGER = logging.getLogger(__name__)
 
 
 class EltakoFlowHandler(config_entries.ConfigFlow, domain=DOMAIN):
-    """Handle the Eltako config flows."""
+    """Simplified Eltako config flow for testing."""
 
     VERSION = 1
     MINOR_VERSION = 1
-    MANUAL_PATH_VALUE = "Custom path"
 
-    def __init__(self) -> None:
-        """Initialize the Eltako config flow."""
-
-    def is_input_available(self, user_input) -> bool:
-        if user_input is not None:
-            if CONF_SERIAL_PATH in user_input and user_input[CONF_SERIAL_PATH] is not None:
-                if CONF_GATEWAY_DESCRIPTION in user_input and user_input[CONF_GATEWAY_DESCRIPTION] is not None:
-                    return True
-        return False
-
-    async def async_step_user(self, user_input=None):
-        """Handle an Eltako config flow start."""
-        # is called when adding a new gateway
-        LOGGER.debug("config_flow user step started.")
-        return await self.async_step_detect()
-
-    async def async_step_detect(self, user_input=None):
-        """Propose a list of detected gateways."""
-        LOGGER.debug("config_flow detect step started.")
-        return await self.manual_selection_routine(user_input)
-        
-    async def async_step_manual(self, user_input=None):
-        """Request manual USB gateway path."""
-        LOGGER.debug("config_flow manual step started.")
-        return await self.manual_selection_routine(user_input, manual_setp=True)
-    
-    async def manual_selection_routine(self, user_input=None, manual_setp:bool=False):
-        LOGGER.debug("Add new gateway")
+    async def async_step_user(
+        self, user_input: dict[str, Any] | None = None
+    ) -> FlowResult:
+        """Handle the initial step."""
         errors = {}
 
-        # get configuration for debug purpose
-        config = await config_helpers.async_get_home_assistant_config(self.hass, CONFIG_SCHEMA)
-        LOGGER.debug(f"Config: {config}\n")
-
-        # ensure data entry is set
-        if DATA_ELTAKO not in self.hass.data:
-            LOGGER.debug("No configuration available.")
-            self.hass.data.setdefault(DATA_ELTAKO, {})
-
-        # goes recursively ...
-        # check if values were set in the step before
         if user_input is not None:
-            if self.is_input_available(user_input):
-                if await self.validate_eltako_conf(user_input):
-                    return self.create_eltako_entry(user_input)
-            
-                errors = {CONF_SERIAL_PATH: ERROR_INVALID_GATEWAY_PATH}
+            # Basic validation
+            gateway_description = user_input.get(CONF_GATEWAY_DESCRIPTION, "")
+            serial_path = user_input.get(CONF_SERIAL_PATH, "")
 
-        # find all existing serial paths
-        serial_paths = await self.hass.async_add_executor_job(gateway.detect)
-        
-        # get available (not registered) gateways
-        g_list_dict = (await config_helpers.async_get_list_of_gateway_descriptions(self.hass, CONFIG_SCHEMA))
-        # filter out registered gateways. all registered gateways are listen in data section
-        g_list = list([g for g in g_list_dict.values() if g not in self.hass.data[DATA_ELTAKO]])
-        LOGGER.debug("Available gateways to be added: %s", g_list)
-        if len(g_list) == 0:
-            LOGGER.debug("No gateways are configured in the 'configuration.yaml'.")
-            errors = {CONF_GATEWAY_DESCRIPTION: ERROR_NO_GATEWAY_CONFIGURATION_AVAILABLE}
+            if not gateway_description:
+                errors[CONF_GATEWAY_DESCRIPTION] = "Gateway description required"
+            if not serial_path:
+                errors[CONF_SERIAL_PATH] = "Serial path required"
 
-        # add manually added serial paths and ip addresses from configuration
-        for g_id in g_list_dict.keys():
-            g_c = config_helpers.find_gateway_config_by_id(config, g_id)
-            if CONF_SERIAL_PATH in g_c:
-                serial_paths.append(g_c[CONF_SERIAL_PATH])
-            if CONF_GATEWAY_ADDRESS in g_c:
-                serial_paths.append(g_c[CONF_GATEWAY_ADDRESS])
+            if not errors:
+                # Create entry - this should trigger dependency installation
+                return self.async_create_entry(
+                    title="Eltako",
+                    data={
+                        CONF_GATEWAY_DESCRIPTION: gateway_description,
+                        CONF_SERIAL_PATH: serial_path,
+                    }
+                )
 
-        # get all serial paths which are not taken by existing gateways
-        device_registry = dr.async_get(self.hass)
-        serial_paths_of_registered_gateways = await gateway.async_get_serial_path_of_registered_gateway(device_registry)
-        serial_paths = list(set([sp for sp in serial_paths if sp not in serial_paths_of_registered_gateways]))
-        LOGGER.debug("Available serial paths/IP addresses: %s", serial_paths)
-
-        if manual_setp or len(serial_paths) == 0:
-            LOGGER.debug("No usb port or any manually configured address available.")
-            errors = {CONF_SERIAL_PATH: ERROR_NO_SERIAL_PATH_AVAILABLE}
-                
-            return self.async_show_form(
-                step_id="manual",
-                data_schema=vol.Schema({
-                    vol.Required(CONF_GATEWAY_DESCRIPTION, msg="EnOcean Gateway", description="Gateway to be initialized."): vol.In(g_list),
-                    vol.Required(CONF_SERIAL_PATH, msg="Serial Port/IP Address", description="Serial path/IP address for selected gateway."): str
-                }),
-                errors=errors,
-            )
-
-
-        # show form in which gateways and serial paths are displayed so that a mapping can be selected.
+        # Show form
         return self.async_show_form(
-            step_id="detect",
+            step_id="user",
             data_schema=vol.Schema({
-                vol.Required(CONF_GATEWAY_DESCRIPTION, msg="EnOcean Gateway", description="Gateway to be initialized."): vol.In(g_list),
-                vol.Required(CONF_SERIAL_PATH, msg="Serial Port/IP Address", description="Serial path/IP address for selected gateway."): vol.In(serial_paths),
+                vol.Required(CONF_GATEWAY_DESCRIPTION, default="Test Gateway (Gateway-0)"): str,
+                vol.Required(CONF_SERIAL_PATH, default="/dev/ttyUSB0"): str,
             }),
             errors=errors,
         )
 
-    async def validate_eltako_conf(self, user_input) -> bool:
-        """Return True if the user_input contains a valid gateway path."""
-        serial_path: str = user_input[CONF_SERIAL_PATH]
-        baud_rate: int = -1
-        gateway_selection: str = user_input[CONF_GATEWAY_DESCRIPTION]
-
-        for gdc in gateway.GatewayDeviceType:
-            if gdc in gateway_selection:
-                baud_rate = gateway.BAUD_RATE_DEVICE_TYPE_MAPPING[gdc]
-                break
-        
-        # check ip address for esp3 over tcp
-        if GatewayDeviceType.LAN in gateway_selection:
-            try:
-                ip = ipaddress.ip_address(serial_path)
-                return True
-            except Exception:
-                return False
-        # check serial ports / usb
-        else:
-            path_is_valid = await self.hass.async_add_executor_job(
-                gateway.validate_path, serial_path, baud_rate
-            )
-            LOGGER.debug("serial_path: %s, validated with baud rate %d is %s", serial_path, baud_rate, path_is_valid)
-
-        return path_is_valid
-
-    def create_eltako_entry(self, user_input):
-        """Create an entry for the provided configuration."""
-        return self.async_create_entry(title="Eltako", data=user_input)
-
     @staticmethod
     @callback
-    def async_get_options_flow(config_entry: config_entries.ConfigEntry) -> EltakoOptionsFlowHandler:
-        """Create the options flow."""
+    def async_get_options_flow(config_entry):
+        """Return the options flow."""
         return EltakoOptionsFlowHandler(config_entry)
 
 
 class EltakoOptionsFlowHandler(config_entries.OptionsFlow):
-    """Handle Eltako options flow."""
+    """Handle Eltako options."""
 
-    def __init__(self, config_entry: config_entries.ConfigEntry) -> None:
+    def __init__(self, config_entry):
         """Initialize options flow."""
         self.config_entry = config_entry
 
-    async def async_step_init(self, user_input: dict[str, Any] | None = None) -> FlowResult:
-        """Handle options flow."""
+    async def async_step_init(self, user_input=None):
+        """Manage the options."""
         if user_input is not None:
             return self.async_create_entry(title="", data=user_input)
-
-        # Get current gateway settings
-        gateway_description = self.config_entry.data.get(CONF_GATEWAY_DESCRIPTION, "")
-        serial_path = self.config_entry.data.get(CONF_SERIAL_PATH, "")
-
-        # Get current options
-        current_options = self.config_entry.options
-        auto_reconnect = current_options.get(CONF_GATEWAY_AUTO_RECONNECT, True)
-        message_delay = current_options.get(CONF_GATEWAY_MESSAGE_DELAY, 0.0)
-        enable_teach_in = current_options.get(CONF_ENABLE_TEACH_IN_BUTTONS, True)
 
         return self.async_show_form(
             step_id="init",
             data_schema=vol.Schema({
-                vol.Optional(
-                    CONF_GATEWAY_AUTO_RECONNECT,
-                    description="Enable automatic reconnection",
-                    default=auto_reconnect
-                ): bool,
-                vol.Optional(
-                    CONF_GATEWAY_MESSAGE_DELAY,
-                    description="Message delay in seconds",
-                    default=message_delay
-                ): vol.Coerce(float),
-                vol.Optional(
-                    CONF_ENABLE_TEACH_IN_BUTTONS,
-                    description="Enable teach-in buttons",
-                    default=enable_teach_in
-                ): bool,
-            }),
-            description_placeholders={
-                "gateway_name": gateway_description,
-                "serial_path": serial_path,
-            },
+                vol.Optional("test_option", default=False): bool,
+            })
         )
